feat: create ScrollableKeyboardView component and add it to layouts
